MTN: Refactored import-from-firehose logic into its own file
